recommender/requirements.txt
pydantic
langchain
transformers
torch
faiss-cpu
pillow
accelerate
langchain-ollama